add: парсинг комментариев

+++ ParserAnimeGO/Models/DubbingAnimeData.cs
@@ -2,7 +2,7 @@
 {
     public class DubbingAnimeData
     {
-        public int IdFromAnimeGo { get; set; }
+        public long IdFromAnimeGo { get; set; }
         public List<string> Dubbing { get; set; } = new List<string>();
     }
 }
